requirements.txt
boto3
click
python-dotenv